refactor(mcp): extract git_guards module from git.py (#299)
Move six guard functions and ten constants from the 754-line
mcp/git.py into a new mcp/git_guards.py; git.py re-exports every
name so existing monkeypatch targets continue to resolve. Remove
xfail markers from W9 RED tests. CHANGELOG entry added.
