+++ CHANGELOG.md
@@ -95,6 +95,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `tracing/tracer`: `get_tracer_from_settings` pins `MERGECRAFT_TRACE_ID` on first mint and reuses a process-wide `Tracer` when tracing settings match, so MCP-style `tools/call` handlers on a worker thread share one Logfire `trace_id` instead of minting a new tree per call (#292)
 - `tracing/_tool_attrs`: `enrich_tool_response` now classifies `ToolResult.is_error=True` (and mapping keys `is_error`/`isError`) as an error span even when no Python exception is raised — `tool.exit_code=error`, span status `error`, and `gen_ai.tool.output` set for the GenAI dashboard; JSON-RPC is unchanged (#296)
 - `analyzers/trust`: `build_review_source` detects a linked git worktree of the same repo (same `git rev-parse --git-common-dir`) and sets kind `"local_worktree"`, which `derive_source_trust_tier` maps to `"trusted"` without `--trust`; temporary clones and unrelated repos remain `"untrusted"` (#294)
+- `mcp/git_guards`: six guard functions (`_is_config_flag`, `_subcommand_declares_shorts`, `_reject_config_flags`, `_reject_namespace_flag`, `_reject_branch_writes`, `_reject_file_writing_flags`) and ten guard constants extracted from the 754-line `mcp/git.py` into a dedicated `mcp/git_guards.py` module; `mcp/git.py` re-exports every name so existing `monkeypatch` targets continue to resolve (#299)
 - `cli/auth`: authenticating a provider left the contributor unable to run the CLI locally — `mergecraft auth codex|claude|gemini|cursor|nous|tokenhub|minimax` persisted the captured credential only through `gh secret set`, so it landed in GitHub Actions secrets and nowhere the local runtime reads. All seven now take `--scope local|github|both` (default `github`, so a CI-only setup keeps today's behaviour exactly, including the hard failure when the secret write fails), and `--scope local` needs no `gh` auth and no network. `auth codex` captures `auth.json` before its isolated `CODEX_HOME` tempdir is deleted, and the pretty-printed multi-line payload `codex login` writes is compacted to one line and quoted, so the `.env` entry round-trips through both `dotenv` and `source .env` (#221)
 - `cli/auth`: running `mergecraft auth … --scope local` from a subdirectory reported success while writing a nested `.env` that nothing reads — the local path fell back to `.env` in the current working directory; it is now anchored at the git repository root, and a run with no repository to anchor to fails with a pointer to `MERGECRAFT_ENV` instead of writing somewhere useless. `$MERGECRAFT_ENV` still wins when set (#221)
 - `agents/codex`: every Codex review failed to start whenever a custom OpenAI-compatible provider was configured (`MERGECRAFT_CUSTOM_PROVIDER_*`) — the generated `config.toml` emitted the `default_permissions` profile name *after* the `[model_providers.*]` tables, so TOML scoped it into a provider block and Codex never saw the read-only review permission profile; the generated file now always emits every root scalar ahead of the first table header, whatever order the writer populated them in, so the same ordering also covers `sandbox_mode` on the non-profile path (#222)
